i dont know

+++ mod/src/game.cpp
@@ -76,7 +76,7 @@ int Game::checkPowerBricks()
 
     int i = -1; // -1 is no powerbrick.
     // how many right shifts // WARN: does not account for loading a different save or losing hostage progress.
-    for (DWORD64 a =  extraPurchasedPrev ^ extraPurchased; a > 0; a = a >> 1) i++;
+    for (DWORD64 a =  extraPurchasedPrev ^ extraPurchased; a > 0; a = a >> ((DWORD64)1)) i++;
     return i; 
 }
